+++ src/timing.ts
@@ -13,8 +13,10 @@ import { CustomToEpochCommand } from './commands/customToEpochCommand';
 import { CustomToIsoLocalCommand } from './commands/customToIsoLocalCommand';
 import { CustomToIsoUtcCommand } from './commands/customToIsoUtcCommand';
 import { EpochToCustomCommand } from './commands/epochToCustomCommand';
+import { EpochToISODurationCommand } from './commands/epochToISODurationCommand';
 import { EpochToIsoLocalCommand } from './commands/epochToIsoLocalCommand';
 import { EpochToIsoUtcCommand } from './commands/epochToIsoUtcCommand';
+import { EpochToReadableDurationCommand } from './commands/epochToReadableDurationCommand';
 import { IsoRfcToCustomCommand } from './commands/isoRfcToCustomCommand';
 import { IsoRfcToEpochCommand } from './commands/isoRfcToEpochCommand';
 import { NowAsCustomCommand } from './commands/nowAsCustomCommand';
@@ -23,8 +25,9 @@ import { NowAsIsoLocalCommand } from './commands/nowAsIsoLocalCommand';
 import { NowAsIsoUtcCommand } from './commands/nowAsIsoUtcCommand';
 import { ToggleInsertConvertedTimeUserLevelCommand } from './commands/toggleInsertConvertedTimeUserLevelCommand';
 
+import { DurationHoverProvider } from './util/durationHoverProvider';
 import { TimeConverter } from './util/timeConverter';
-import { TimeHoverProvider } from './util/timeHoverProvider';
+import { TimestampHoverProvider } from './util/timestampHoverProvider';
 
 export function activate(context: vscode.ExtensionContext) {
 
@@ -36,6 +39,8 @@ export function activate(context: vscode.ExtensionContext) {
     const customToIsoUtcCommand = new CustomToIsoUtcCommand(context, timeConverter);
     const customToIsoLocalCommand = new CustomToIsoLocalCommand(context, timeConverter);
     const epochToCustomCommand = new EpochToCustomCommand(context, timeConverter);
+    const epochToReadableDurationCommand = new EpochToReadableDurationCommand(context, timeConverter);
+    const epochToIsoDurationCommand = new EpochToISODurationCommand(context, timeConverter);
     const epochToIsoLocalCommand = new EpochToIsoLocalCommand(context, timeConverter);
     const epochToIsoUtcCommand = new EpochToIsoUtcCommand(context, timeConverter);
     const nowAsEpochCommand = new NowAsEpochCommand(context, timeConverter);
@@ -47,7 +52,8 @@ export function activate(context: vscode.ExtensionContext) {
 
     const toggleInsertConvertedTimeUserLevelCommand = new ToggleInsertConvertedTimeUserLevelCommand();
 
-    const timeHoverProvider = new TimeHoverProvider(timeConverter);
+    const timestampHoverProvider = new TimestampHoverProvider(timeConverter);
+    const durationHoverProvider = new DurationHoverProvider(timeConverter);
     /* tslint:disable:max-line-length */
 
     context.subscriptions.push(
@@ -56,6 +62,8 @@ export function activate(context: vscode.ExtensionContext) {
         customToIsoLocalCommand, vscode.commands.registerCommand('timing.customToIsoLocal', customToIsoLocalCommand.execute, customToIsoLocalCommand),
         customToIsoUtcCommand, vscode.commands.registerCommand('timing.customToIsoUtc', customToIsoUtcCommand.execute, customToIsoUtcCommand),
         epochToCustomCommand, vscode.commands.registerCommand('timing.epochToCustom', epochToCustomCommand.execute, epochToCustomCommand),
+        epochToReadableDurationCommand, vscode.commands.registerCommand('timing.epochToReadableDuration', epochToReadableDurationCommand.execute, epochToReadableDurationCommand),
+        epochToIsoDurationCommand, vscode.commands.registerCommand('timing.epochToIsoDuration', epochToIsoDurationCommand.execute, epochToIsoDurationCommand),
         epochToIsoLocalCommand, vscode.commands.registerCommand('timing.epochToIsoLocal', epochToIsoLocalCommand.execute, epochToIsoLocalCommand),
         epochToIsoUtcCommand, vscode.commands.registerCommand('timing.epochToIsoUtc', epochToIsoUtcCommand.execute, epochToIsoUtcCommand),
         isoRfcToCustomCommand, vscode.commands.registerCommand('timing.isoRfcToCustom', isoRfcToCustomCommand.execute, isoRfcToCustomCommand),
@@ -67,7 +75,8 @@ export function activate(context: vscode.ExtensionContext) {
 
         vscode.commands.registerCommand('timing.toggleInsertConvertedTimeUserLevel', toggleInsertConvertedTimeUserLevelCommand.execute, toggleInsertConvertedTimeUserLevelCommand),
         // Register Hover Provider
-        timeHoverProvider, vscode.languages.registerHoverProvider('*', timeHoverProvider)
+        timestampHoverProvider, vscode.languages.registerHoverProvider('*', timestampHoverProvider),
+        durationHoverProvider, vscode.languages.registerHoverProvider('*', durationHoverProvider)
     );
     /* tslint:enable:max-line-length */
 }